requirements.txt
bcrypt
fuzzywuzzy
pycrypto
pyperclip
python-Levenshtein

